requirements.txt
# Dependências do projeto GeoChat

# Para comunicação assíncrona com RabbitMQ
pika>=1.3.0

# Para carregar variáveis de ambiente
python-dotenv>=1.0.0

# Para interface gráfica (já vem com Python)
# tkinter - built-in

# Para cálculos matemáticos (já vem com Python)  
# math - built-in

# Para threading e sockets (já vem com Python)
# threading - built-in
# socket - built-in

# Para JSON (já vem com Python)
# json - built-in

# Para manipulação de datas (já vem com Python)
# datetime - built-in
